[EDIT] - reword party component with 3 components (add-track, playlist & player)
