+++ src/helpers/route-constant.js
@@ -16,4 +16,3 @@ export const MYINVOICES = "/myinvoices";
 export const MYPARTIES = "/parties";
 export const PARTYSHOW = "/party/:slug";
 export const PARTYEDIT = "/party/:slug/edit";
-export const TRACKNEW = "/party/:slug/track/new";
